+++ Reposotory.Implement/MonthlyRentRepository.cs
@@ -115,7 +115,7 @@ namespace Reposotory.Implement
             checkSQLResult(ref flag, SPOutput.Error, SPOutput.ErrorCode, ref lstError, ref errCode);
             return flag;
         }
-        public bool InsMonthlyHistory(string IDNO, Int64 OrderNo,Int64 MonthlyRentId, float UseWorkDayHours,float UseHolidayHours, float UseMotoTotalHours, Int64 LogID, ref string errCode)
+        public bool InsMonthlyHistory(string IDNO, Int64 OrderNo,Int64 MonthlyRentId, int UseWorkDayHours, int UseHolidayHours, int UseMotoTotalHours, Int64 LogID, ref string errCode)
         {
             bool flag = false;
             List<ErrorInfo> lstError = new List<ErrorInfo>();
@@ -172,6 +172,5 @@ namespace Reposotory.Implement
                 }
             }
         }
-#endregion
     }
 }
